test(green): 등록되지 않은 닉네임일 때 예외 처리 테스트
