+++ README.md
@@ -29,6 +29,7 @@
 ### 출석 확인 기능
 
 - [ ] 닉네임과 등교 시간을 입력받는다.
+    - [x] 등록되지 않은 닉네임의 경우 예외 처리한다.
 - [x] 출석 후 출석 기록을 확인할 수 있다.
 - [x] 이미 출석한 경우, 다시 출석할 수 없으며 수정 기능을 이용하도록 안내한다.
     - [x] 출석하는 날짜에 해당하는 크루의 출석 데이터가 있는지 확인한다.
