+++ src/test/java/CrewTest.java
@@ -1,7 +1,5 @@
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.AttendanceRecord;
-import domain.DateProvider;
 import java.time.LocalDate;
 import org.junit.jupiter.api.Test;
 
@@ -15,18 +13,8 @@ public class CrewTest {
         Crew crew = new Crew(nickname, time,
                 () -> LocalDate.of(2024, 12, 13));
 
-        assertThat(crew.nickname).isEqualTo(nickname);
+        assertThat(crew.getNickname()).isEqualTo(nickname);
     }
 
-    class Crew {
 
-        private final String nickname;
-        private final AttendanceRecord attendanceRecord;
-
-        public Crew(String nickname, String time, DateProvider dateProvider) {
-            this.nickname = nickname;
-            this.attendanceRecord = new AttendanceRecord(dateProvider);
-            this.attendanceRecord.attend(time);
-        }
-    }
 }
